(PC-13519)[PRO] refactor: create 2 different components for Offer route
