+++ pro/src/routes/Offers/CollectiveOffers/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CollectiveOffers'
