+++ pro/src/routes/Offers/IndividualOffers/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './IndividualOffers'
