+++ pro/src/routes/Offers/Offers.tsx
@@ -1,70 +1,23 @@
-import React, { useCallback, useEffect, useState } from 'react'
-import { useDispatch } from 'react-redux'
-import { useHistory } from 'react-router-dom'
+import React, { useEffect, useState } from 'react'
 
 import useActiveFeature from 'components/hooks/useActiveFeature'
-import useCurrentUser from 'components/hooks/useCurrentUser'
 import useNotification from 'components/hooks/useNotification'
-import Spinner from 'components/layout/Spinner'
-import { computeOffersUrl } from 'components/pages/Offers/utils/computeOffersUrl'
-import { DEFAULT_PAGE, DEFAULT_SEARCH_FILTERS } from 'core/Offers/constants'
+import { DEFAULT_SEARCH_FILTERS } from 'core/Offers'
 import { useQuerySearchFilters } from 'core/Offers/hooks'
-import { Audience, Offer, Offerer, TSearchFilters } from 'core/Offers/types'
-import OffersScreen from 'screens/Offers'
-import { savePageNumber, saveSearchFilters } from 'store/offers/actions'
+import { Audience, Offerer } from 'core/Offers/types'
 
-import { getFilteredOffersAdapter, getOffererAdapter } from './adapters'
+import { getOffererAdapter } from './adapters'
+import CollectiveOffers from './CollectiveOffers'
+import IndividualOffers from './IndividualOffers'
 
 const Offers = (): JSX.Element => {
-  const history = useHistory()
-  const notify = useNotification()
   const [urlSearchFilters, urlPageNumber, urlAudience] = useQuerySearchFilters()
-  const { currentUser } = useCurrentUser()
-  const dispatch = useDispatch()
+  const notify = useNotification()
+
+  const [offerer, setOfferer] = useState<Offerer | null>(null)
 
   const separateIndividualAndCollectiveOffers = useActiveFeature(
     'ENABLE_INDIVIDUAL_AND_COLLECTIVE_OFFER_SEPARATION'
-  )
-
-  const [offerer, setOfferer] = useState<Offerer | null>(null)
-  const [offers, setOffers] = useState<Offer[]>([])
-  const [isLoading, setIsLoading] = useState(true)
-  const [initialSearchFilters, setInitialSearchFilters] =
-    useState<TSearchFilters | null>(null)
-
-  const loadAndUpdateOffers = useCallback(
-    async (filters: TSearchFilters) => {
-      const apiFilters = {
-        ...DEFAULT_SEARCH_FILTERS,
-        ...filters,
-      }
-      const { isOk, message, payload } = await getFilteredOffersAdapter(
-        apiFilters
-      )
-
-      if (!isOk) {
-        setIsLoading(false)
-        return notify.error(message)
-      }
-
-      setIsLoading(false)
-      setOffers(payload.offers)
-    },
-    [notify]
-  )
-
-  const redirectWithUrlFilters = useCallback(
-    (
-      filters: TSearchFilters & {
-        page?: number
-        audience?: Audience
-      }
-    ) => {
-      const newUrl = computeOffersUrl(filters, filters.page)
-
-      history.push(newUrl)
-    },
-    [history]
   )
 
   useEffect(() => {
@@ -88,70 +41,24 @@ const Offers = (): JSX.Element => {
     loadOfferer()
   }, [urlSearchFilters.offererId, notify])
 
-  useEffect(() => {
-    const filters = { ...DEFAULT_SEARCH_FILTERS, ...urlSearchFilters }
-    if (currentUser.isAdmin) {
-      const isVenueFilterSelected =
-        urlSearchFilters.venueId !== DEFAULT_SEARCH_FILTERS.venueId
-      const isOffererFilterApplied =
-        urlSearchFilters.offererId !== DEFAULT_SEARCH_FILTERS.offererId
-      const isFilterByVenueOrOfferer =
-        isVenueFilterSelected || isOffererFilterApplied
-
-      if (!isFilterByVenueOrOfferer) {
-        filters.status = DEFAULT_SEARCH_FILTERS.status
-      }
-    }
-    setInitialSearchFilters(filters)
-  }, [setInitialSearchFilters, urlSearchFilters, currentUser.isAdmin])
-
-  useEffect(() => {
-    dispatch(
-      saveSearchFilters({
-        nameOrIsbn:
-          urlSearchFilters.nameOrIsbn || DEFAULT_SEARCH_FILTERS.nameOrIsbn,
-        offererId:
-          urlSearchFilters.offererId || DEFAULT_SEARCH_FILTERS.offererId,
-        venueId: urlSearchFilters.venueId || DEFAULT_SEARCH_FILTERS.venueId,
-        categoryId:
-          urlSearchFilters.categoryId || DEFAULT_SEARCH_FILTERS.categoryId,
-        status: urlSearchFilters.status
-          ? urlSearchFilters.status
-          : DEFAULT_SEARCH_FILTERS.status,
-        creationMode: urlSearchFilters.creationMode
-          ? urlSearchFilters.creationMode
-          : DEFAULT_SEARCH_FILTERS.creationMode,
-        periodBeginningDate:
-          urlSearchFilters.periodBeginningDate ||
-          DEFAULT_SEARCH_FILTERS.periodBeginningDate,
-        periodEndingDate:
-          urlSearchFilters.periodEndingDate ||
-          DEFAULT_SEARCH_FILTERS.periodEndingDate,
-      })
-    )
-    dispatch(savePageNumber(urlPageNumber))
-  }, [dispatch, urlPageNumber, urlSearchFilters])
-
-  if (!initialSearchFilters) {
-    return <Spinner />
-  }
-
-  return (
-    <OffersScreen
-      currentPageNumber={urlPageNumber ?? DEFAULT_PAGE}
-      currentUser={currentUser}
-      initialSearchFilters={initialSearchFilters}
-      isLoading={isLoading}
-      loadAndUpdateOffers={loadAndUpdateOffers}
+  return urlAudience === Audience.INDIVIDUAL ? (
+    <IndividualOffers
       offerer={offerer}
-      offers={offers}
-      redirectWithUrlFilters={redirectWithUrlFilters}
       separateIndividualAndCollectiveOffers={
         separateIndividualAndCollectiveOffers
       }
-      setIsLoading={setIsLoading}
       setOfferer={setOfferer}
-      urlAudience={urlAudience}
+      urlPageNumber={urlPageNumber}
+      urlSearchFilters={urlSearchFilters}
+    />
+  ) : (
+    <CollectiveOffers
+      offerer={offerer}
+      separateIndividualAndCollectiveOffers={
+        separateIndividualAndCollectiveOffers
+      }
+      setOfferer={setOfferer}
+      urlPageNumber={urlPageNumber}
       urlSearchFilters={urlSearchFilters}
     />
   )
